+++ src/Akkatecture/Extensions/TypeExtensions.cs
@@ -1,23 +1,22 @@
 ﻿// The MIT License (MIT)
 //
-// Copyright (c) 2015-2020 Rasmus Mikkelsen
-// Copyright (c) 2015-2020 eBay Software Foundation
-// Modified from original source https://github.com/eventflow/EventFlow
+// Copyright (c) 2015-2021 Rasmus Mikkelsen
+// Copyright (c) 2015-2021 eBay Software Foundation
+//     Modified from original source https://github.com/eventflow/EventFlow
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors
+//    https://github.com/AfterLutz/Akketecture
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture 
-// 
-// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
@@ -43,7 +42,7 @@ namespace Akkatecture.Extensions
 {
     public static class TypeExtensions
     {
-        private static readonly ConcurrentDictionary<Type, string> PrettyPrintCache = new ConcurrentDictionary<Type, string>();
+        private static readonly ConcurrentDictionary<Type, string> PrettyPrintCache = new();
 
         public static string PrettyPrint(this Type type)
         {
@@ -65,15 +64,11 @@ namespace Akkatecture.Extensions
         private static string PrettyPrintRecursive(Type type, int depth)
         {
             if (depth > 3)
-            {
                 return type.Name;
-            }
 
             var nameParts = type.Name.Split('`');
             if (nameParts.Length == 1)
-            {
                 return nameParts[0];
-            }
 
             var genericArguments = type.GetTypeInfo().GetGenericArguments();
             return !type.IsConstructedGenericType
@@ -81,7 +76,7 @@ namespace Akkatecture.Extensions
                 : $"{nameParts[0]}<{string.Join(",", genericArguments.Select(t => PrettyPrintRecursive(t, depth + 1)))}>";
         }
 
-        private static readonly ConcurrentDictionary<Type, AggregateName> AggregateNames = new ConcurrentDictionary<Type, AggregateName>();
+        private static readonly ConcurrentDictionary<Type, AggregateName> AggregateNames = new();
 
         public static AggregateName GetAggregateName(
             this Type aggregateType)
@@ -90,10 +85,8 @@ namespace Akkatecture.Extensions
                 aggregateType,
                 t =>
                 {
-                    if (!typeof(IAggregateRoot).GetTypeInfo().IsAssignableFrom(aggregateType))
-                    {
-                        throw new ArgumentException($"Type '{aggregateType.PrettyPrint()}' is not an aggregate root");
-                    }
+                    if (!typeof(IAggregateRoot).GetTypeInfo().IsAssignableFrom(t))
+                        throw new ArgumentException($"Type '{t.PrettyPrint()}' is not an aggregate root");
 
                     return new AggregateName(
                         t.GetTypeInfo().GetCustomAttributes<AggregateNameAttribute>().SingleOrDefault()?.Name ??
@@ -101,8 +94,7 @@ namespace Akkatecture.Extensions
                 });
         }
 
-
-        private static readonly ConcurrentDictionary<Type, AggregateName> SagaNames = new ConcurrentDictionary<Type, AggregateName>();
+        private static readonly ConcurrentDictionary<Type, AggregateName> SagaNames = new();
 
         public static AggregateName GetSagaName(
             this Type sagaType)
@@ -111,18 +103,16 @@ namespace Akkatecture.Extensions
                 sagaType,
                 t =>
                 {
-                    if (!typeof(IAggregateRoot).GetTypeInfo().IsAssignableFrom(sagaType))
-                    {
-                        throw new ArgumentException($"Type '{sagaType.PrettyPrint()}' is not a saga.");
-                    }
+                    if (!typeof(IAggregateRoot).GetTypeInfo().IsAssignableFrom(t))
+                        throw new ArgumentException($"Type '{t.PrettyPrint()}' is not a saga.");
 
                     return new AggregateName(
                         t.GetTypeInfo().GetCustomAttributes<SagaNameAttribute>().SingleOrDefault()?.Name ??
                         t.Name);
                 });
         }
-        
-        private static readonly ConcurrentDictionary<Type, JobName> JobNames = new ConcurrentDictionary<Type, JobName>();
+
+        private static readonly ConcurrentDictionary<Type, JobName> JobNames = new();
 
         public static JobName GetJobName(
             this Type jobType)
@@ -131,10 +121,8 @@ namespace Akkatecture.Extensions
                 jobType,
                 t =>
                 {
-                    if (!typeof(IJob).GetTypeInfo().IsAssignableFrom(jobType))
-                    {
-                        throw new ArgumentException($"Type '{jobType.PrettyPrint()}' is not a job");
-                    }
+                    if (!typeof(IJob).GetTypeInfo().IsAssignableFrom(t))
+                        throw new ArgumentException($"Type '{t.PrettyPrint()}' is not a job");
 
                     return new JobName(
                         t.GetTypeInfo().GetCustomAttributes<JobNameAttribute>().SingleOrDefault()?.Name ??
@@ -163,7 +151,7 @@ namespace Akkatecture.Extensions
                     mi => mi.GetParameters()[0].ParameterType,
                     mi => ReflectionHelper.CompileMethodInvocation<Action<T, IAggregateEvent>>(type, "Apply", mi.GetParameters()[0].ParameterType));
         }
-        
+
         internal static IReadOnlyDictionary<Type, Action<T, IAggregateSnapshot>> GetAggregateSnapshotHydrateMethods<TAggregate, TIdentity, T>(this Type type)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity
@@ -216,15 +204,15 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
+
             var domainEventTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ISubscribeToAsync<,,>))
-                .Select(i =>   typeof(IDomainEvent<,,>).MakeGenericType(i.GetGenericArguments()[0],i.GetGenericArguments()[1],i.GetGenericArguments()[2]))
+                .Select(i => typeof(IDomainEvent<,,>).MakeGenericType(i.GetGenericArguments()[0],i.GetGenericArguments()[1],i.GetGenericArguments()[2]))
                 .ToList();
-            
 
             return domainEventTypes;
         }
-        
+
         internal static IReadOnlyList<Type> GetAggregateExecuteTypes(this Type type)
         {
             var interfaces = type
@@ -232,11 +220,11 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
+
             var domainEventTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IExecute<>))
-                .Select(i =>   i.GetGenericArguments()[0])
+                .Select(i => i.GetGenericArguments()[0])
                 .ToList();
-            
 
             return domainEventTypes;
         }
@@ -248,23 +236,24 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
+
             var domainEventTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ISubscribeTo<,,>))
-                .Select(i =>   typeof(IDomainEvent<,,>).MakeGenericType(i.GetGenericArguments()[0],i.GetGenericArguments()[1],i.GetGenericArguments()[2]))
+                .Select(i => typeof(IDomainEvent<,,>).MakeGenericType(i.GetGenericArguments()[0],i.GetGenericArguments()[1],i.GetGenericArguments()[2]))
                 .ToList();
-            
 
             return domainEventTypes;
         }
 
-        private static readonly ConcurrentDictionary<Type, AggregateName> AggregateNameCache = new ConcurrentDictionary<Type, AggregateName>();
+        private static readonly ConcurrentDictionary<Type, AggregateName> AggregateNameCache = new();
+
         internal static AggregateName GetCommittedEventAggregateRootName(this Type type)
         {
             return AggregateNameCache.GetOrAdd(
                 type,
                 t =>
                 {
-                    var interfaces = type
+                    var interfaces = t
                         .GetTypeInfo()
                         .GetInterfaces()
                         .Select(i => i.GetTypeInfo())
@@ -274,22 +263,22 @@ namespace Akkatecture.Extensions
                         .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ICommittedEvent<,>))
                         .Select(i => i.GetGenericArguments()[0]).SingleOrDefault();
 
-
                     if (aggregateType != null)
                         return aggregateType.GetAggregateName();
 
-                    throw new ArgumentException(nameof(type));
+                    throw new ArgumentException(nameof(t));
                 });
         }
-        
-        private static readonly ConcurrentDictionary<Type, Type> AggregateEventTypeCache = new ConcurrentDictionary<Type, Type>();
+
+        private static readonly ConcurrentDictionary<Type, Type> AggregateEventTypeCache = new();
+
         internal static Type GetCommittedEventAggregateEventType(this Type type)
         {
             return AggregateEventTypeCache.GetOrAdd(
                 type,
                 t =>
                 {
-                    var interfaces = type
+                    var interfaces = t
                         .GetTypeInfo()
                         .GetInterfaces()
                         .Select(i => i.GetTypeInfo())
@@ -299,14 +288,13 @@ namespace Akkatecture.Extensions
                         .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ICommittedEvent<,,>))
                         .Select(i => i.GetGenericArguments()[2]).SingleOrDefault();
 
-
                     if (aggregateEvent != null)
                         return aggregateEvent;
 
-                    throw new ArgumentException(nameof(type));
+                    throw new ArgumentException(nameof(t));
                 });
         }
-        
+
         internal static IReadOnlyList<Type> GetJobRunTypes(this Type type)
         {
             var interfaces = type
@@ -314,15 +302,15 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
+
             var jobRunTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IRun<>))
-                .Select(i =>   i.GetGenericArguments()[0])
+                .Select(i => i.GetGenericArguments()[0])
                 .ToList();
-            
 
             return jobRunTypes;
         }
-        
+
         internal static IReadOnlyList<Type> GetAsyncSagaEventSubscriptionTypes(this Type type)
         {
             var interfaces = type
@@ -342,12 +330,12 @@ namespace Akkatecture.Extensions
                 .Select(t => typeof(IDomainEvent<,,>).MakeGenericType(t.GetGenericArguments()[0],
                     t.GetGenericArguments()[1], t.GetGenericArguments()[2]))
                 .ToList();
-            
+
             startedByEventTypes.AddRange(handleEventTypes);
 
             return startedByEventTypes;
         }
-        
+
         internal static IReadOnlyList<Type> GetSagaEventSubscriptionTypes(this Type type)
         {
             var interfaces = type
@@ -367,12 +355,12 @@ namespace Akkatecture.Extensions
                 .Select(t => typeof(IDomainEvent<,,>).MakeGenericType(t.GetGenericArguments()[0],
                     t.GetGenericArguments()[1], t.GetGenericArguments()[2]))
                 .ToList();
-            
+
             startedByEventTypes.AddRange(handleEventTypes);
 
             return startedByEventTypes;
         }
-        
+
         internal static IReadOnlyList<Type> GetAsyncSagaTimeoutSubscriptionTypes(this Type type)
         {
             var interfaces = type
@@ -380,14 +368,15 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
-            
+
             var sagaTimeoutSubscriptionTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ISagaHandlesTimeoutAsync<>))
                 .Select(t => t.GetGenericArguments()[0])
                 .ToList();
-            
+
             return sagaTimeoutSubscriptionTypes;
         }
+
         internal static IReadOnlyList<Type> GetSagaTimeoutSubscriptionTypes(this Type type)
         {
             var interfaces = type
@@ -395,40 +384,41 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
-            
+
             var sagaTimeoutSubscriptionTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(ISagaHandlesTimeout<>))
                 .Select(t => t.GetGenericArguments()[0])
                 .ToList();
-            
+
             return sagaTimeoutSubscriptionTypes;
         }
-        
+
         internal static IReadOnlyDictionary<Type, Func<T,IAggregateEvent, IAggregateEvent>> GetAggregateEventUpcastMethods<TAggregate, TIdentity, T>(this Type type)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity
         {
             var aggregateEventType = typeof(IAggregateEvent<TAggregate, TIdentity>);
-            
+
             return type
                 .GetTypeInfo()
                 .GetMethods(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
                 .Where(mi =>
                 {
-                    if (mi.Name != "Upcast") 
+                    if (mi.Name != "Upcast")
                         return false;
+
                     var parameters = mi.GetParameters();
                     return
                         parameters.Length == 1 &&
                         aggregateEventType.GetTypeInfo().IsAssignableFrom(parameters[0].ParameterType);
-                    
+
                 })
                 .ToDictionary(
                     //problem might be here
                     mi => mi.GetParameters()[0].ParameterType,
-                    mi => ReflectionHelper.CompileMethodInvocation<Func<T,IAggregateEvent, IAggregateEvent>>(type, "Upcast", mi.GetParameters()[0].ParameterType));           
+                    mi => ReflectionHelper.CompileMethodInvocation<Func<T,IAggregateEvent, IAggregateEvent>>(type, "Upcast", mi.GetParameters()[0].ParameterType));
         }
-        
+
         internal static IReadOnlyList<Type> GetAggregateEventUpcastTypes(this Type type)
         {
             var interfaces = type
@@ -436,7 +426,7 @@ namespace Akkatecture.Extensions
                 .GetInterfaces()
                 .Select(i => i.GetTypeInfo())
                 .ToList();
-            
+
             var upcastableEventTypes = interfaces
                 .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IUpcast<,>))
                 .Select(i =>   i.GetGenericArguments()[0])
@@ -452,11 +442,11 @@ namespace Akkatecture.Extensions
             while (currentType != null)
             {
                 if (currentType.Name.Contains(name))
-                {
                     return currentType;
-                }
+
                 currentType = currentType.BaseType;
             }
+
             return type;
         }
     }