+++ src/Akkatecture/Jobs/JobScheduler.cs
@@ -3,8 +3,9 @@
 // Copyright (c) 2017-2020 Bartosz Sypytkowski
 // Modified from original source https://github.com/Horusiath/Akka.Persistence.Reminders
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture 
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors  
+//    https://github.com/AfterLutz/Akketecture
 // 
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of