+++ src/Akkatecture/Sagas/AggregateSaga/AggregateSaga.cs
@@ -1,11 +1,12 @@
 // The MIT License (MIT)
 //
-// Copyright (c) 2015-2020 Rasmus Mikkelsen
-// Copyright (c) 2015-2020 eBay Software Foundation
+// Copyright (c) 2015-2021 Rasmus Mikkelsen
+// Copyright (c) 2015-2021 eBay Software Foundation
 // Modified from original source https://github.com/eventflow/EventFlow
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors  
+//    https://github.com/AfterLutz/Akketecture
 //
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -53,7 +54,7 @@ namespace Akkatecture.Sagas.AggregateSaga
         private static readonly IReadOnlyDictionary<Type, Action<TSagaState, IAggregateEvent>> ApplyMethodsFromState = typeof(TSagaState).GetAggregateStateEventApplyMethods<TAggregateSaga, TIdentity, TSagaState>();
         private static readonly IReadOnlyDictionary<Type, Action<TSagaState, IAggregateSnapshot>> HydrateMethodsFromState = typeof(TSagaState).GetAggregateSnapshotHydrateMethods<TAggregateSaga, TIdentity, TSagaState>();
         private static readonly IAggregateName SagaName = typeof(TAggregateSaga).GetSagaName();
-        private static readonly List<Type> _sagaTimeoutTypes = new List<Type>();
+        private readonly List<Type> _sagaTimeoutTypes = new List<Type>();
         private Dictionary<Type, IActorRef> SagaTimeoutManagers { get; set; }
         private CircularBuffer<ISourceId> _previousSourceIds = new CircularBuffer<ISourceId>(100);
         
@@ -107,8 +108,11 @@ namespace Akkatecture.Sagas.AggregateSaga
                 InitAsyncReceives();
             }
 
-            InitTimeoutJobManagers();
-            InitAsyncTimeoutJobManagers();
+            if (Settings.UseSagaTimeouts)
+            {
+                InitTimeoutJobManagers();
+                InitAsyncTimeoutJobManagers();
+            }
 
             if (Settings.UseDefaultEventRecover)
             {
@@ -136,7 +140,7 @@ namespace Akkatecture.Sagas.AggregateSaga
                 var sagaTimeoutManagerType = typeof(SagaTimeoutManager<>).MakeGenericType(sagaTimeoutType);
                 var sagaTimeoutManager = Context.ActorOf(Props.Create(() => 
                         (ActorBase) Activator.CreateInstance(sagaTimeoutManagerType)), 
-                    $"{sagaTimeoutType.Name}-timeoutmanager");
+                    $"{sagaTimeoutType.Name.ToLower()}-timeoutmanager");
                 SagaTimeoutManagers.Add(sagaTimeoutType, sagaTimeoutManager);
             }
         }