+++ test/Akkatecture.NodeTestRunner/Sink.cs
@@ -4,10 +4,12 @@
 // Copyright (c) 2013 - 2020 .NET Foundation
 // Modified from original source https://github.com/akkadotnet/akka.net
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture 
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors  
+//    https://github.com/AfterLutz/Akketecture
 // 
-// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// 
+// Permission is hereby granted free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
@@ -107,4 +109,3 @@ namespace Akkatecture.NodeTestRunner
         }
     }
 }
-