+++ test/Akkatecture.TestHelpers/Aggregates/Commands/CreateTestCommand.cs
@@ -1,7 +1,8 @@
 ﻿// The MIT License (MIT)
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture 
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors  
+//    https://github.com/AfterLutz/Akketecture
 // 
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -30,6 +31,14 @@ namespace Akkatecture.TestHelpers.Aggregates.Commands
         public CreateTestCommand(
             TestAggregateId aggregateId,
             CommandId sourceId)
+            : base(aggregateId, sourceId)
+        {
+        }
+    }
+
+    public class CreateTestCommandRequestingCommandResult : Command<TestAggregate, TestAggregateId>
+    {
+        public CreateTestCommandRequestingCommandResult(TestAggregateId aggregateId, CommandId sourceId)
             : base(aggregateId, sourceId)
         {
         }