+++ test/Akkatecture.TestHelpers/Aggregates/TestAggregate.cs
@@ -1,7 +1,8 @@
 ﻿// The MIT License (MIT)
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors  
+//    https://github.com/AfterLutz/Akketecture
 //
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -25,6 +26,7 @@ using System;
 using System.Linq;
 using Akka.Persistence;
 using Akkatecture.Aggregates;
+using Akkatecture.Aggregates.CommandResults;
 using Akkatecture.Aggregates.ExecutionResults;
 using Akkatecture.Aggregates.Snapshot;
 using Akkatecture.Aggregates.Snapshot.Strategies;
@@ -41,6 +43,7 @@ namespace Akkatecture.TestHelpers.Aggregates
     [AggregateName("Test")]
     public sealed class TestAggregate : AggregateRoot<TestAggregate, TestAggregateId, TestAggregateState>, 
         IExecute<CreateTestCommand>,
+        IExecute<CreateTestCommandRequestingCommandResult>,
         IExecute<CreateAndAddTwoTestsCommand>,
         IExecute<AddTestCommand>,
         IExecute<AddFourTestsCommand>,
@@ -80,7 +83,22 @@ namespace Akkatecture.TestHelpers.Aggregates
 
             return true;
         }
-        
+        public bool Execute(CreateTestCommandRequestingCommandResult command)
+        {
+            if (IsNew)
+            {
+                Emit(new TestCreatedEvent(command.AggregateId), new Metadata {{"some-key","some-value"}});
+                Reply(CommandResult.SucceedWith(command));
+            }
+            else
+            {
+                TestErrors++;
+                Throw(new TestedErrorEvent(TestErrors));
+                ReplyFailure(CommandResult.FailWith(command, "Aggregate already exists"));
+            }
+
+            return true;
+        }
 
         public bool Execute(CreateAndAddTwoTestsCommand command)
         {