+++ test/Akkatecture.Tests/IntegrationTests/Aggregates/Sagas/AggregateSagaTests.cs
@@ -1,7 +1,8 @@
 ﻿// The MIT License (MIT)
 //
-// Copyright (c) 2018 - 2020 Lutando Ngqakaza
-// https://github.com/Lutando/Akkatecture 
+// Copyright (c) 2018 - 2021 Lutando Ngqakaza
+// Copyright (c) 2022-2023 AfterLutz Contributors  
+//    https://github.com/AfterLutz/Akketecture
 // 
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -53,11 +54,14 @@ namespace Akkatecture.Tests.IntegrationTests.Aggregates.Sagas
         [Category(Category)]
         public void SendingTest_FromTestAggregate_CompletesSaga()
         {
-            var eventProbe = CreateTestProbe("event-probe");
-            Sys.EventStream.Subscribe(eventProbe, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaStartedEvent>));
-            Sys.EventStream.Subscribe(eventProbe, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaCompletedEvent>));
-            Sys.EventStream.Subscribe(eventProbe, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTransactionCompletedEvent>));
-            Sys.EventStream.Subscribe(eventProbe, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>));
+            var eventProbe1 = CreateTestProbe("event-probe1");
+            var eventProbe2 = CreateTestProbe("event-probe2");
+            var eventProbe3 = CreateTestProbe("event-probe3");
+            var eventProbe4 = CreateTestProbe("event-probe4");
+            Sys.EventStream.Subscribe(eventProbe1, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaStartedEvent>));
+            Sys.EventStream.Subscribe(eventProbe2, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaCompletedEvent>));
+            Sys.EventStream.Subscribe(eventProbe3, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTransactionCompletedEvent>));
+            Sys.EventStream.Subscribe(eventProbe4, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>));
             
             var aggregateManager = Sys.ActorOf(Props.Create(() => new TestAggregateManager()), "test-aggregatemanager");
             Sys.ActorOf(Props.Create(() => new TestSagaManager(() => new TestSaga(aggregateManager))), "test-sagaaggregatemanager");
@@ -78,27 +82,104 @@ namespace Akkatecture.Tests.IntegrationTests.Aggregates.Sagas
             var sagaStartingCommand = new GiveTestCommand(senderAggregateId, CommandId.New,receiverAggregateId,senderTest);
             aggregateManager.Tell(sagaStartingCommand);
             
-            eventProbe.
+            eventProbe1.
                 ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaStartedEvent>>(
                     x => x.AggregateEvent.Sender.Equals(senderAggregateId)
                          && x.AggregateEvent.Receiver.Equals(receiverAggregateId)
                          && x.AggregateEvent.SentTest.Equals(senderTest),new TimeSpan(0,0,20));
             
-            eventProbe.
+            eventProbe3.
                 ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTransactionCompletedEvent>>(new TimeSpan(0,0,20));
             
-            eventProbe.
+            eventProbe2.
                 ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaCompletedEvent>>(new TimeSpan(0,0,20));
                         
-            eventProbe.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
-                timeoutMsg => timeoutMsg.AggregateEvent.TimeoutMessage.Equals("First timeout test"),
+            eventProbe4.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
+                timeoutMsg => timeoutMsg.AggregateEvent.TimeoutMessage.StartsWith("First timeout test"),
                 TimeSpan.FromSeconds(15));
             
-            eventProbe.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
+            eventProbe4.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
                 timeoutMsg => timeoutMsg.AggregateEvent.TimeoutMessage.StartsWith("Second timeout test"),
                 TimeSpan.FromSeconds(15));
         }
-        
+
+        [Fact]
+        [Category(Category)]
+        public void SendingTests_FromTwoTestAggregates_AllowsForTwoSagasWithTimeouts()
+        {
+            var eventProbe1 = CreateTestProbe("event-probe1");
+            Sys.EventStream.Subscribe(eventProbe1, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>));
+            
+            var eventProbe2 = CreateTestProbe("event-probe2");
+            Sys.EventStream.Subscribe(eventProbe2, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>)); 
+            
+            var eventProbe3 = CreateTestProbe("event-probe3");
+            Sys.EventStream.Subscribe(eventProbe3, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>));
+            
+            var eventProbe4 = CreateTestProbe("event-probe4");
+            Sys.EventStream.Subscribe(eventProbe4, typeof(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>));
+            
+            
+            var aggregateManager = Sys.ActorOf(Props.Create(() => new TestAggregateManager()), "test-aggregatemanager");
+            Sys.ActorOf(Props.Create(() => new TestSagaManager(() => new TestSaga(aggregateManager))), "test-sagaaggregatemanager");
+           
+            // Create first sender aggregate
+            var firstSenderId = new TestAggregateId("testaggregate-11111111-1111-1111-1111-111111111111");
+            aggregateManager.Tell(new CreateTestCommand(firstSenderId, CommandId.New));
+
+            // Create second sender aggregate
+            var secondSenderId = new TestAggregateId("testaggregate-22222222-2222-2222-2222-222222222222");
+            aggregateManager.Tell(new CreateTestCommand(secondSenderId, CommandId.New));
+
+            // Create receiver aggregate
+            var receiverAggregateId = TestAggregateId.New;
+            aggregateManager.Tell(new CreateTestCommand(receiverAggregateId, CommandId.New));
+
+            // Create first test entity
+            var firstTestId = TestId.New;
+            var firstTest = new Test(firstTestId);
+            var nextAggregateCommand = new AddTestCommand(firstSenderId, CommandId.New, firstTest);
+            aggregateManager.Tell(nextAggregateCommand);
+            
+            // Send GiveTestCommand to start first saga
+            var sagaStartingCommand = new GiveTestCommand(firstSenderId, CommandId.New,receiverAggregateId,firstTest);
+            aggregateManager.Tell(sagaStartingCommand);
+            
+            // Create second test aggregate
+            var secondTestId = TestId.New;
+            var secondTest = new Test(secondTestId);
+            nextAggregateCommand = new AddTestCommand(secondSenderId, CommandId.New, secondTest);
+            aggregateManager.Tell(nextAggregateCommand);
+            
+            // Send GiveTestCommand to start second saga
+            sagaStartingCommand = new GiveTestCommand(secondSenderId, CommandId.New,receiverAggregateId,secondTest);
+            aggregateManager.Tell(sagaStartingCommand);
+
+            eventProbe1.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
+                timeoutMsg => IsTimeoutMessage(timeoutMsg), 
+                TimeSpan.FromSeconds(15));
+            
+            eventProbe2.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
+                timeoutMsg => IsTimeoutMessage(timeoutMsg),
+                TimeSpan.FromSeconds(15));
+            
+            eventProbe3.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
+                timeoutMsg =>  IsTimeoutMessage(timeoutMsg),
+                TimeSpan.FromSeconds(15));
+            
+            eventProbe4.ExpectMsg<DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred>>(
+                timeoutMsg =>   IsTimeoutMessage(timeoutMsg),
+                TimeSpan.FromSeconds(15));
+        }
+
+        private bool IsTimeoutMessage(DomainEvent<TestSaga, TestSagaId, TestSagaTimeoutOccurred> timeoutMsg)
+        {
+            var containsCorrectSagaIdInMessage = timeoutMsg.AggregateEvent.TimeoutMessage.Contains(timeoutMsg.AggregateIdentity.Value);
+            return containsCorrectSagaIdInMessage && (timeoutMsg.AggregateEvent.TimeoutMessage.StartsWith("First timeout test") || 
+                   timeoutMsg.AggregateEvent.TimeoutMessage.StartsWith("Second timeout test"));
+
+        }
+
         [Fact]
         [Category(Category)]
         public void SendingTest_FromTestAggregate_CompletesSagaAsync()